+++ src/http/middleware/errors/errors-handler.middleware.ts
@@ -1,6 +1,6 @@
 import { Next, Context } from "koa";
 
-import { logger } from "../../logger/logger";
+import { logger } from "../../../logger/logger";
 
 export const errorMiddleware = () => {
     return async (ctx: Context, next: Next) => {
