+++ src/http/middleware/logger/logger.middleware.ts
@@ -1,5 +1,5 @@
 import { Next, Context } from "koa";
-import { logger } from "../../logger/logger";
+import { logger } from "../../../logger/logger";
 import { Packet, Socket } from "socket.io";
 
 export const loggerMiddleware = () => {
@@ -12,9 +12,3 @@ export const loggerMiddleware = () => {
         ctx.status >= 400 ? logger.error(message) : logger.info(message);
     };
 }
-
-export const socketLoggerMiddleware = (socket: Socket) =>
-    ([event, data]: Packet, next: (err?: any) => void) => {
-        logger.info(`socket ${socket.id} on event ${event} received packet: ${JSON.stringify(data)}`);
-        next();
-    }
