+++ src/web-socket/web-socket.ts
@@ -1,18 +1,8 @@
-import { Socket } from "socket.io";
-import * as cookie from 'cookie';
-
 import { webSocket } from "..";
 import { Product } from "../models/product.model";
 import { findProductQuery } from "../database/product.queries";
 import { logger } from "../logger/logger";
-import { addCart, removeCart, calculateProductLimit, removeCartProduct, updateCartProduct, getCart } from "../cart/cart";
-
-export const registerSocket = (socket: Socket, next: (err?: Error) => void) => {
-    socket.id = cookie.parse(socket.request.headers.cookie)['userId.sig'];
-    logger.info(`A user connected ${socket.id}`);
-    addCart(socket.id);
-    next();
-}
+import { removeCart, calculateProductLimit, removeCartProduct, updateCartProduct, getCart } from "../cart/cart";
 
 export const clearCart = async (socketId: string) =>
     await Promise.all(Object.keys(getCart(socketId))
